Fix: Implementation erros and improve types

+++ include/vec_string.h
@@ -1,14 +1,15 @@
 #ifndef VEC_STRING_H
 #define VEC_STRING_H
+#include <stddef.h>
 
 typedef struct {
     char *data;
-    int capacity;
-    int size;
+    size_t capacity;
+    size_t size;
 } String;
 
 
 String* new();
-String* with_capacity(unsigned int capacity);
+String* with_capacity(size_t capacity);
 void delete(String str);
 #endif //VEC_STRING_H
